+++ src/scene-d/HA/HAView/HAViewInformation/LocalComponents/HAInformation.js
@@ -5,10 +5,11 @@ import { Dialog } from "primereact/dialog";
 import React, { useRef, useState } from "react";
 import EmbeddedHelp from "../../../../../app/common/EmbeddedHelp/EmbeddedHelp";
 import FDate from "../../../../../app/common/FDate/FDate";
+import HAStatus from "../../../../../app/common/HAStatus/HAStatus";
 import PredictedDateEditor from "../../../../../app/common/PredictedDateEditor/PredictedDateEditor";
-import TagGeneral from "../../../../../app/common/TagGeneral/TagGeneral";
 
-const HAStatus = ({ project }) => {
+const HAInformation = ({ project }) => {
+  console.log("HAInformation.js: project: ", project);
   const cm = useRef(null);
   const [displayEditContainer, setDisplayEditContainer] = useState(false);
 
@@ -30,19 +31,7 @@ const HAStatus = ({ project }) => {
   let data = [
     {
       name: "HA Status",
-      value: (
-        <TagGeneral
-          tag={
-            project.status === "Active"
-              ? project.currentStage === "HA"
-                ? "Ongoing"
-                : "Complete"
-              : project.currentStage === "HA"
-              ? "Terminated at HA"
-              : "Complete"
-          }
-        />
-      ),
+      value: <HAStatus id={project.id} status={project?.haStatus} />,
     },
     {
       name: "HA Start Date",
@@ -90,4 +79,4 @@ const HAStatus = ({ project }) => {
   );
 };
 
-export default HAStatus;
+export default HAInformation;